Break log proxy into own namespace
